consensus: reduce complexity and handle errors
This patch is a start towards a simpler and more correct API for the
backend teams to lean on.  The consensus service is now a concrete type
that the ApiServer will maintain a reference to.  It by default provides
synchronous methods in which to interact with the underlying KV
database.

Additonally, we'll include methods that will call these synchronous
methods followed by a user provided callback allowing for them to be
handled on a separate go-routine.  For one-off API requests this isn't
too useful, but when we begin to lean on websockets more heavily, the
backend teams will likely need a way to remain responsive to new ws
requests without waiting for DB interactions to complete.