backend: complete restructuring of consensus mod
This patch introduces a ConsensusService and some associated methods
to synchronously communicate with dragonboat storage.  Additionally,
methods in the API server are simpler, as there are rich error values
produced from the ConsensusService.  Some methods are now removed, such
as CallUpdatePixel, as it's functionality can be implemented directly in
the websocket processing.  Note that the only time we (officially) need
to authenticate a user id is when updating a pixel.

Additionally, I've added some examples using sirupsen's logrus logging
package rather than default logging in go.  It's fairly convenient and
is more flexible and much easier to unit test.