ci: backend and frontend actions for build n test
This patch introduces two CI jobs: one to build and deploy our frontend
(static site) to gh-pages and one to build our backend code.  This will
prevent patches that break compilation, but we do not run (any) tests on
either codebase.  Additionally, I don't believe anything is cached when
building the backend, so CI runs are going to be really fucking long...